ページを削除するメソッドの実装を追加

+++ app/backend/src/ScoreHistoryApi/Logics/IScoreDatabase.cs
@@ -64,7 +64,7 @@ namespace ScoreHistoryApi.Logics
         /// <param name="ownerId"></param>
         /// <param name="scoreId"></param>
         /// <param name="pageIds"></param>
-        Task RemovePagesAsync(Guid ownerId, Guid scoreId, List<int> pageIds);
+        Task RemovePagesAsync(Guid ownerId, Guid scoreId, List<long> pageIds);
 
         /// <summary>
         /// ページの置き換え
